fix: move pages styles to separate folder
