+++ pages-styles/Board.styled.ts
