+++ pages/404/index.tsx
@@ -1,5 +1,5 @@
 import type { NextPage } from 'next';
-import { NotFoundMain, NotFoundContent } from './404.styled.page';
+import { NotFoundMain, NotFoundContent } from '../../pages-styles/404.styled';
 
 const NotFound: NextPage = () => {
   return (
